Changes Read to receive uint64 byte count. (#2085)

+++ imports/assemblyscript/assemblyscript.go
@@ -305,7 +305,7 @@ func readAssemblyScriptString(mem api.Memory, offset uint32) (string, bool) {
 	if !ok || byteCount%2 != 0 {
 		return "", false
 	}
-	buf, ok := mem.Read(offset, byteCount)
+	buf, ok := mem.Read(offset, uint64(byteCount))
 	if !ok {
 		return "", false
 	}
