+++ imports/wasi_snapshot_preview1/random.go
@@ -41,7 +41,7 @@ func randomGetFn(_ context.Context, mod api.Module, params []uint64) sys.Errno {
 	randSource := sysCtx.RandSource()
 	buf, bufLen := uint32(params[0]), uint32(params[1])
 
-	randomBytes, ok := mod.Memory().Read(buf, bufLen)
+	randomBytes, ok := mod.Memory().Read(buf, uint64(bufLen))
 	if !ok { // out-of-range
 		return sys.EFAULT
 	}
